+++ src/PairsManager.py
@@ -1,6 +1,5 @@
 # region imports
 from AlgorithmImports import *
-from src.constants import PositionMode
 from typing import Dict, Set
 # endregion